Move to typer package
gage-typer was misguided. Can't just rename a package without renaming
the packaged modules!

Some funny business with the `incompatible_with` callback - the change
here is made without understanding. Without this change we get an
internal error in typer as it gets a surprise None for the param
value. My hunch is that the callback is being called with the untyped
value. If that's by design, this fix is correct. That's a surprising
design though as returning an unmodified value causes type related
errors downstream.

This is not a huge issue though either way. This callback never worked
very well to begin with. We're probably better off moving this
outside the type declaration and into the impl. Or typer might have
added something like this (I haven't looked lately)