Normalize write status for user, group, and other
Coerce group and other permission to reflect user permission across
systems. This removes the variability that can occur across systems
wrt default group and other write permissions.

The tests still cover the expected behavior, which is to differentiate
between read-only and writeable files.

A less opporunistic (hacky) approach would be to report a simple "read
only vs writeable" status for the user and print that. These changes
to the equivalent but by faking out writeable status for group and
other, as per user write status.